Adicionado programa que verifica se os número recebidos estão em ordem

+++ buzz.py
@@ -1,4 +1,4 @@
-numero = float(input("Qual número deseja saber se é divisível por 5? "))
+numero = int(input("Qual número deseja saber se é divisível por 5? "))
 if (numero % 5 == 0):
     print("Buzz")
 else:
