+++ fizz.py
@@ -1,4 +1,4 @@
-numero = float(input("Qual número deseja saber se é divisível por 3? "))
+numero = int(input("Qual número deseja saber se é divisível por 3? "))
 if (numero % 3 == 0):
     print("Fizz")
 else:
