+++ paridade.py
@@ -1,4 +1,4 @@
-numero = float(input("Qual número deseja saber se é par ou ímpar? "))
+numero = int(input("Qual número deseja saber se é par ou ímpar? "))
 if (numero % 2 == 0):
     print("par")
 else:
